Remove coupling between pkl.core.module/resource and pkl.core.messaging
