+++ pkl-cli/src/main/kotlin/org/pkl/cli/CliTestRunner.kt
@@ -17,9 +17,9 @@ package org.pkl.cli
 
 import java.io.Writer
 import org.pkl.commons.cli.*
+import org.pkl.core.Closeables
 import org.pkl.core.EvaluatorBuilder
 import org.pkl.core.ModuleSource.uri
-import org.pkl.core.Closeables
 import org.pkl.core.stdlib.test.report.JUnitReport
 import org.pkl.core.stdlib.test.report.SimpleReport
 import org.pkl.core.util.ErrorMessages
