+++ pkl-core/src/main/java/org/pkl/core/messaging/Messages.java
@@ -20,17 +20,20 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import org.pkl.core.messaging.Message.*;
+import org.pkl.core.module.ExternalModuleResolver;
 import org.pkl.core.module.PathElement;
+import org.pkl.core.resource.ExternalResourceResolver;
 import org.pkl.core.util.Nullable;
 
 public final class Messages {
   private Messages() {}
 
   public record ModuleReaderSpec(
-      String scheme, boolean hasHierarchicalUris, boolean isLocal, boolean isGlobbable) {}
+      String scheme, boolean hasHierarchicalUris, boolean isLocal, boolean isGlobbable)
+      implements ExternalModuleResolver.Spec {}
 
-  public record ResourceReaderSpec(
-      String scheme, boolean hasHierarchicalUris, boolean isGlobbable) {}
+  public record ResourceReaderSpec(String scheme, boolean hasHierarchicalUris, boolean isGlobbable)
+      implements ExternalResourceResolver.Spec {}
 
   public record ListResourcesRequest(long requestId, long evaluatorId, URI uri)
       implements Server.Request {
