+++ pkl-core/src/main/java/org/pkl/core/resource/ExternalResourceResolver.java
@@ -18,113 +18,26 @@ package org.pkl.core.resource;
 import java.io.IOException;
 import java.net.URI;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.Random;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.Future;
 import org.pkl.core.SecurityManager;
 import org.pkl.core.SecurityManagerException;
-import org.pkl.core.messaging.MessageTransport;
-import org.pkl.core.messaging.MessageTransports;
-import org.pkl.core.messaging.Messages.*;
-import org.pkl.core.messaging.ProtocolException;
 import org.pkl.core.module.PathElement;
 
-public class ExternalResourceResolver {
-  private final MessageTransport transport;
-  private final long evaluatorId;
-  private final Map<URI, Future<byte[]>> readResponses = new ConcurrentHashMap<>();
-  private final Map<URI, Future<List<PathElement>>> listResponses = new ConcurrentHashMap<>();
-  private final Random requestIdGenerator = new Random();
+public interface ExternalResourceResolver {
 
-  public ExternalResourceResolver(MessageTransport transport, long evaluatorId) {
-    this.transport = transport;
-    this.evaluatorId = evaluatorId;
+  interface Spec {
+    boolean hasHierarchicalUris();
+
+    boolean isGlobbable();
+
+    String scheme();
   }
 
-  public Optional<Object> read(URI uri) throws IOException {
-    var result = doRead(uri);
-    return Optional.of(new Resource(uri, result));
-  }
+  Optional<Object> read(URI uri) throws IOException;
 
-  public boolean hasElement(SecurityManager securityManager, URI elementUri)
-      throws SecurityManagerException {
-    securityManager.checkResolveResource(elementUri);
-    try {
-      doRead(elementUri);
-      return true;
-    } catch (IOException e) {
-      return false;
-    }
-  }
+  boolean hasElement(SecurityManager securityManager, URI elementUri)
+      throws SecurityManagerException;
 
-  public List<PathElement> listElements(SecurityManager securityManager, URI baseUri)
-      throws IOException, SecurityManagerException {
-    securityManager.checkResolveResource(baseUri);
-    return doListElements(baseUri);
-  }
-
-  public List<PathElement> doListElements(URI baseUri) throws IOException {
-    return MessageTransports.resolveFuture(
-        listResponses.computeIfAbsent(
-            baseUri,
-            (uri) -> {
-              var future = new CompletableFuture<List<PathElement>>();
-              var request =
-                  new ListResourcesRequest(requestIdGenerator.nextLong(), evaluatorId, uri);
-              try {
-                transport.send(
-                    request,
-                    (response) -> {
-                      if (response instanceof ListResourcesResponse resp) {
-                        if (resp.error() != null) {
-                          future.completeExceptionally(new IOException(resp.error()));
-                        } else {
-                          future.complete(
-                              Objects.requireNonNullElseGet(resp.pathElements(), List::of));
-                        }
-                      } else {
-                        future.completeExceptionally(new ProtocolException("unexpected response"));
-                      }
-                    });
-              } catch (ProtocolException | IOException e) {
-                future.completeExceptionally(e);
-              }
-              return future;
-            }));
-  }
-
-  public byte[] doRead(URI baseUri) throws IOException {
-    return MessageTransports.resolveFuture(
-        readResponses.computeIfAbsent(
-            baseUri,
-            (uri) -> {
-              var future = new CompletableFuture<byte[]>();
-              var request =
-                  new ReadResourceRequest(requestIdGenerator.nextLong(), evaluatorId, uri);
-              try {
-                transport.send(
-                    request,
-                    (response) -> {
-                      if (response instanceof ReadResourceResponse resp) {
-                        if (resp.error() != null) {
-                          future.completeExceptionally(new IOException(resp.error()));
-                        } else if (resp.contents() != null) {
-                          future.complete(resp.contents());
-                        } else {
-                          future.complete(new byte[0]);
-                        }
-                      } else {
-                        future.completeExceptionally(new ProtocolException("unexpected response"));
-                      }
-                    });
-              } catch (ProtocolException | IOException e) {
-                future.completeExceptionally(e);
-              }
-              return future;
-            }));
-  }
+  List<PathElement> listElements(SecurityManager securityManager, URI baseUri)
+      throws IOException, SecurityManagerException;
 }
