+++ pkl-core/src/main/java/org/pkl/core/resource/ResourceReaders.java
@@ -31,7 +31,6 @@ import org.pkl.core.SecurityManager;
 import org.pkl.core.SecurityManagerException;
 import org.pkl.core.externalreader.ExternalReaderProcess;
 import org.pkl.core.externalreader.ExternalReaderProcessException;
-import org.pkl.core.messaging.Messages.ResourceReaderSpec;
 import org.pkl.core.module.FileResolver;
 import org.pkl.core.module.ModulePathResolver;
 import org.pkl.core.module.PathElement;
@@ -163,7 +162,7 @@ public final class ResourceReaders {
 
   /** Returns a reader for external and client reader resources. */
   public static ResourceReader externalResolver(
-      ResourceReaderSpec spec, ExternalResourceResolver resolver) {
+      ExternalResourceResolver.Spec spec, ExternalResourceResolver resolver) {
     return new ExternalResolver(spec, resolver);
   }
 
@@ -692,10 +691,11 @@ public final class ResourceReaders {
   }
 
   private static final class ExternalResolver implements ResourceReader {
-    private final ResourceReaderSpec readerSpec;
+    private final ExternalResourceResolver.Spec readerSpec;
     private final ExternalResourceResolver resolver;
 
-    public ExternalResolver(ResourceReaderSpec readerSpec, ExternalResourceResolver resolver) {
+    public ExternalResolver(
+        ExternalResourceResolver.Spec readerSpec, ExternalResourceResolver resolver) {
       this.readerSpec = readerSpec;
       this.resolver = resolver;
     }
@@ -721,7 +721,7 @@ public final class ResourceReaders {
     }
 
     @Override
-    public boolean hasElement(org.pkl.core.SecurityManager securityManager, URI elementUri)
+    public boolean hasElement(SecurityManager securityManager, URI elementUri)
         throws SecurityManagerException {
       return resolver.hasElement(securityManager, elementUri);
     }
